# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      code:
        example: 1001
        type: integer
      error:
        example: Internal technical details
        type: string
      message:
        example: User-friendly message
        type: string
    type: object
...
  dto.OTPRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dto.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.VerifyOTPRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
...
host: identity-provider.isaxbsit2027.com
info:
  contact:
    email: causonmikolorenz@gmail.com
    name: Miko Lorenz Causon
  description: Identity Provider for the Unified School System Capstone.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Unified Access Identity Provider API
  version: "1.0"
paths:
...
  /mail/otp:
    post:
      consumes:
      - application/json
      description: Generates a 6-digit numeric OTP and sends it via email.
      parameters:
      - description: OTP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Send OTP Code
      tags:
      - mail
...
  /otp/send:
    post:
      consumes:
      - application/json
      description: Generates a 6-digit numeric OTP and sends it via email.
      parameters:
      - description: OTP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Send OTP Code
      tags:
      - otp
  /otp/verify:
    post:
      consumes:
      - application/json
      description: Verifies a 6-digit numeric OTP for a user.
      parameters:
      - description: Verify OTP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Verify OTP Code
      tags:
      - otp
...
securityDefinitions:
  ApiKeyAuth:
    description: Your custom API Key for identification.
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: Type 'Bearer ' followed by your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"